UNST-5753 Addition of unused output items in proc_def modifies other output items - the differences seem to be caused by round-off differences in EXTALG = EXTALG + BIOMAS*EXTCF
Former-commit-id: 0ebc0f31e4b74b0f00d903b73f6f740a1163acb5

+++ src/engines_gpl/waq/packages/waq_process/src/waq_process/extina.f
@@ -35,7 +35,7 @@
 !
       INTEGER  NALG  , ISWFIX, NIPALG, IFLUX , ISEG  ,
      +         IALG  , IP    , IFIX
-      REAL     EXTALG, EXTCF , BIOMAS
+      REAL(8)  EXTALG, EXTCF , BIOMAS
 !
       NALG  = NINT(PMSA(IPOINT(1)))
       ISWFIX= NINT(PMSA(IPOINT(2)))
